+++ jest.config.js
@@ -3,7 +3,7 @@ module.exports = {
   testEnvironment: 'node',
   roots: ['<rootDir>/src', '<rootDir>/tests'],
   testMatch: ['**/__tests__/**/*.ts', '**/?(*.)+(spec|test).ts'],
-  testPathIgnorePatterns: ['/node_modules/', '/tests/parity/'],
+  testPathIgnorePatterns: ['/node_modules/', '/tests/parity/', '/tests/performance/'],
   collectCoverageFrom: [
     'src/**/*.ts',
     '!src/**/*.d.ts',